refactor(auto-tstm): extract apply helper for CodeScene limit
Move commitAutoTstmPreview into autoTstmApply.ts to keep
useAutoTstmActions under the large-method threshold and document
isDialogFooter for DeepSource.
